[+FEATURE] Fluid (Core): {settings} is available in Partials
Now, the {settings} are automatically available in partials and sections.
Before, they had to be passed explicitely, i.e. by calling 
<f:render section="..." arguments="{settings: settings}" />.

If somebody defined his own "settings"-argument, this still takes precedence:
<f:render section="..." arguments="{settings: some.different.settings}" />
This means that this change is completely backwards compatible.

NOTE: The settings are NOT merged together, like it has been suggested in #6289,
as this will lead to un-obvious behavior for the user.

The test case for this change can be found in I414a82f7beb0e13a615ed39acde93d3c3546ccc1
so make sure to apply these two patches together.

Change-Id: I14c6dd4d9db13a98f1873a79bb43fbf299d2c835
Resolves: #6289