[!!!][FEATURE] Compilation of Fluid templates to PHP files
This adds a compile step to the parsing process of templates, layouts and
partials.
If no compiled cache file exists, Fluid will create it while parsing the
respective Fluid template.
The cache file is invalidated as soon as the template/layout/partial
has been modified (or if caches are flushed).
This speeds up the rendering process by factor 2-5!

Breaking change:

* AbstractViewHelper::$arguments are no longer an object but an array.
If you used $this->arguments->hasArgument() in your custom ViewHelpers,
you'll have to replace this with $this->hasArgument()

Change-Id: Ia1e1ccd364a1d966b6c840273351ab00410c55cf
Resolves: #28544