[FEATURE] (ViewHelpers): Render FlashMessage title and severity
Currently the FlashMessages ViewHelper only renders the body of
the respective Flash Messages.
This change modifies the ViewHelper in a backwards compatible
way so that it outputs the message title if specified.
Additionally the severity (notice, warning, error, ok) is
added to each list item.

Additionally free rendering foreach style is possible now.

Change-Id: Icf2c3e0bff3140bfd7da0c3c0808cca0097f6197
Resolves: #31019
Release: 1.1