[FEATURE] GroupedForViewHelper accepts property path as groupBy argument
This change allows the GroupedForViewHelper to accept not only a property/array key as
argument but also a property path. This reduces memory consumption for more complex
object graphs. Also, It enables a developer to create a custom algorithm (getter) to provide
the value to compare. The change is backwards compatible.

Related: #9244
Releases: 1.1
Change-Id: I09bdf4b6392540df05e1802068c87dd467dcb030