[BUGFIX] AbstractFormFieldViewHelper always converts entities
AbstractFormFieldViewHelper::getValue() currently always tries
to convert objects to identities by calling
PersistenceManager::getIdentifierByObject().
This makes it impossible for custom ViewHelpers to access the
unmodified value that was previously submitted (or bound to the
form field view helper).

This change introduces an argument $convertObjects that allows
experienced developers to disable this behavior in custom
view helpers.

Change-Id: I5326c9424cfecf7a78d9a0ea3c8044150f35bfe4
Fixes: #33569
Releases: 1.1