[!!!][TASK] Move link.widget & uri.widget Widget subfolder
This moves widget related ViewHelpers link.widget & uri.widget
to the Widget subfolder (as in v4).

Instead of link.widget you should use widget.link, instead of
uri.widget widget.uri from now on.

This is not a breaking change right now as the old classes
are still there (with @deprecated annotation). But if you
used one of the ViewHelpers you should adjust your code.

Change-Id: If1db3d17073890e84aa0dd88083bd3eaf8dc9ee5
Resolves: #34231
Releases: 1.1