[BUGFIX] Fix request handling for Ajax widgets
Requests to AJAX widgets were not handled properly
because the routing configuration was not available
to the request handler being in charge for handling
that request.

This patch removes the specialized WidgetRequestHandler
completely and implements the specialized handling of
AJAX widget requests through an aspect which advices
the original Router.

This change set also introduces functional tests which
cover the basic widget request handling executed through
the full request / response stack.

Change-Id: I93da33920f37f1fe38121600f2da890346ffc37f
Resolves: #36500
Resolves: #36236
Releases: 1.1