[!!!][FEATURE] Allow ResourceViewHelper to accept resource:// paths
This enables the path argument of the Uri.ResourceViewHelper
to accept also resource URIs in the format "resource://Package.Name/Public/...".
This functionality used to be achievable with the former uri
argument but was removed with I92dccba6b5acd623ff33eb538e62d0682f00b95e.

This is marked breaking because it introduces Exceptions in case
of incorrect argument usage or values.

Change-Id: I349e18cad4dfda2c1bf16d65e8b12646956e48fc
Releases: 1.2