[!!!][BUGFIX] Mark action argument required in routing related ViewHelpers
Currently you can omit the action attribute when creating a link/URI in
Fluid. If no action was specified the default action of the matching route
was assumed. With #43589 the behavior changed and you'll have to specify
all defaults of a route.

This change makes the action argument of follwing ViewHelpers required:

* f:link.action
* f:uri.action
* f:form

This is a breaking change if you created a link without specifying the action.
In this case change your Fluid code from
<f:action> to <f:action action="index"> (or whatever action is the default
of the target controller).

Change-Id: I03fa30be2522ee53d5a8f184b57afe539ed4f991
Related: #43589
Releases: 1.2