[BUGFIX] Allow one of action and actionUri to be omitted in Form ViewHelper
With I03fa30be2522ee53d5a8f184b57afe539ed4f991 the action argument
of the Form ViewHelper was made required. But in fact you can override the
action attribute by specifying the "actionUri" argument. In that case the
"action" argument should not be required.

This removes this requirement again. But if neither "action" nor "actionUri"
is specified, an exception will be thrown.

This also removes the obsolete method setFormActionUri()

Change-Id: Ifa4ccaafb550526ec977d93059ca123b18ef5462
Related: #43589
Releases: 1.2