[!!!][TASK] Clean up placeholder on textfield and text area VHs
The form.textfield VH had a placeholder attribute as argument to the
render() method; the form.textarea VH did not support placeholder at all.

With this change the placeholder is available on both VHs but registered
as tag attribute.

If you extended TextfieldViewHelper::render() make sure to remove the
$placeholder argument, otherwise you will see

 Argument "placeholder" has already been defined, thus it should not
 be defined again.

errors from Fluid.

Change-Id: I24bb73c5de5a494259171ff289d434892f85e16e
Resolves: #37558
Releases: master