[FEATURE] Configurable prepend option for Select ViewHelper
A common use case with select fields is to have a 'empty' option at
first position in the options list.
This change provides the possibility to add a label and (optionally)
a value, which will be rendered in the option list.

Usage:
<f:form.select prependOptionLabel="select one" />

Note: The prepended option can be translated using the "translate"
argument.

Change-Id: I8d249d6cc7f0c802a53e85d54329f827cb82aba5
Resolves: #44235
Releases: master