[BUGFIX] Resolve duplicate layout and partial compilations
Layouts and partials are independent of the current controller and hence should
not have the controller name in their identifier to prevent duplicate
compilations and cache instances of the same content.
This change removes the controller name from identifiers of layouts and
partials, but leaves the identifier of templates untouched.

Change-Id: I4f6e5b576d9e17e15a3e16cb581bfa4edc54637a
Fixes: #39428
Releases: master, 2.2