[TASK] IfAccessViewHelper now avoids exceptions in Security.log
Previously this view helper was just calling decideOnResource() in
order to find out if a resource was accessible and caught any exception
thrown by that method. Now the correct method hasAccessToResource() is
called instead.

Because the logging aspect registered even caught exceptions, any call
to this view helper resulted in an Access Denied entry in Security.log

Change-Id: I39abae46be9e553954e3967ecabe229124acb551
Releases: master