[BUGFIX] Make required work on Textfield and Textarea VH
The way required was checked any non-null value would mark a field as
required - including FALSE, which is not expected behavior.

This change fixes this by strictly comparing to TRUE, before marking as
required, thus following the documented argument type.

The TextareaViewHelper also gains support for the required attribute
with this change.

Change-Id: I579fbcdc4ff0d1cd5ad14bc9f7560288dc3a1335
Releases: master, 2.2