[!!!][FEATURE] Throw exception for unresolved namespaces
With this change the Fluid parser now throws an exception when it comes
across an unknown ViewHelper namespace.

That is especially helpful if you forgot to import a namespace or
mistyped a ViewHelper name.

It is a breaking change if you rely on the previous behavior of
ignoring ViewHelpers with unknown namespaces.
In that case you can ignore all unknown namespaces with::

  {namespace *}

Specific namespaces can be ignored like this::

  {namespace xs*}  <!-- ignores namespaces starting with "xs" -->
  {namespace foo}  <!-- ignores the namespace "foo" -->

Releases: master
Resolves: FLOW-150
Change-Id: I9c63e7d0b116f9d1baabf09c6c18e97d8deb83fa