[FEATURE] Allow custom CLDR formatting for localized date
This feature adds the ``cldrFormat`` argument to the date format
ViewHelper. If given it will take precedence over the
``localeFormatType`` and ``localeFormatLength`` arguments.
It accepts a date formatting string according to CLDR rules, that
are found here: http://cldr.unicode.org/translation/date-time

Without this feature it is not possible to get custom localized
date strings.

Change-Id: I5b4846bd2675beac31054979fc480eb5944a6e58
Releases: master
Resolves: FLOW-31