[BUGFIX] Adjust code migration identifier pattern to contain the full timestamp
Previously code migrations are expected to have a class name with the
pattern ``Version<YYYYMMDDhhmm>`` and the unique identifier was
determined extracting the last 12 characters of the class name (which
are expected to be the timestamp).

This change adjusts existing code migrations to use the full timestamp in
order to establish the new guideline (note: those migrations still
return the old identifier so that they won't be applied again with a
new identifier).

Change-Id: Ie1b52c6da90ee6c59fb5fa0b644ebe5833b7ec85
Related: FLOW-110
Releases: master, 2.3, 2.2, 2.1, 2.0