[BUGFIX] Always respect property binding in hidden fields
This is a follow up to the
``Submitted form data has precedence over value argument`` change
(I80e7e664e7a1fa41dc36bdf89e331086c6815f78) that had the nasty side
effect of breaking hidden fields bound to properties of a view model.

Before this fix, a template like::

  <f:form.hidden property="someProperty" />

Would lead to a hidden field with an empty ``value`` attribute.

Background:
The custom implementation of the ``getValueAttribute()`` in the
``HiddenViewHelper`` was missing the evaluation of
``getPropertyValue()``. This is now fixed for all of the form
ViewHelpers by making ``AbstractFormFieldViewHelper::getValueAttribute()``
more generic. The bug only really occurred for hidden fields though
because checkbox and radio fields require the ``value`` argument to be
specified explicitly.

Change-Id: Ie78ba551d0a2cade8fe2082db5960f25ee6a9e52
Related: FLOW-213
Releases: master, 3.0