[BUGFIX] Fix support for multi-part namespace identifiers
This change fixes support for namespaces containing dots.
Previously the ``TemplateParser`` skipped those when used in "shorthand
syntax" and something like::

 {foo.bar:baz()}

was never replaced.

Background:
This limitation is not new but it was easy to work around in previous
versions. Since Flow package keys are registered as namespace
identifiers (see FLOW-151) this issue became more urgent.

Change-Id: I5bb99e3ecba1f62e8506b050a429da3d737d31c2
Releases: master, 3.0
Fixes: FLOW-228