[TASK] Get rid of CompilableInterface
From now on, any ViewHelper is considered compilable. A default compile() method is implemented on the AbstractViewHelper class and the ViewHelperInvoker is used as default invokation method. Any ViewHelper can override the compile() method and/or the renderStatic() method to further improve the performance once cached.

ViewHelpers which explicitly must disable compiling for whichever reason, must now do so via the compile() method to set the appropriate flag on the ParsingState, preventing the TemplateParser from attempting to store the template in which the ViewHelper is used.

Also includes a range of performance optimisations for compiled templates, among other things removing further steps in how BooleanNode's stack is compiled; creating an array instead of using serialize() which may be vulnerable on some template storages.