[TASK] Handle layout name directly from variable container
Rather than analyse a whole stack of nodes and compile those to return this single value, the value is now pulled directly from the variablecontainer.

This also means that dynamic layout names can be achieved without touching Fluid itself - simply assign a `layoutName` variable to the view before rendering and it the value of it will be used as layout name.