[TASK] Add method in compiled template to add namespaces
This method is required to (re-)add namespaces which were defined in the template itself rather than added via the ViewHelperResolver. After this change, an additional method is called on each compiled template to add namespaces which were detected during parsing. The compiled namespaces are stored as a simple array.