[BUGFIX] Resolving of Partials, Layout when identifier includes format
Before this change, Fluid would simply append the desired View format to any path that one tried to resolve, which meant it was impossible to request a specific format when rendering a partial by including the file extension. This patch makes Fluid check for the file both with and without the expected view's format file extension.