[FEATURE] Allow compiler to be disabled through API
When the compiler is set to disabled, any template that is attempted stored will be automatically flushed. This allows the compiler to be explicitly disabled by special ViewHelpers - for example, those that *must* have access to child nodes on every execution and depends on postParseEvent integration without having compiling support.

Although it is an edge case it's vital to support these since getting rid of the CompilableInterface which used to trigger this behavior when not detected on a ViewHelper instance. It also makes sense in a development context when a developer wants to prevent compiling while creating a ViewHelper.

An ideal place to trigger this is from within ExpressionNodes (one will be created which can deactivate the compiler) and ViewHelpers' compile() method, both of which have immediate access to the TemplateCompiler/TemplateParser instances via the rendering context.