[BUGFIX] Correct compiling of complex boolean nodes
Okay, so, this patch affects quite a lot of pieces of code but now passes every use case and unit test regarding use of complex and simple conditions alike. The compiling of boolean nodes is changed so it now uses a compiled array of expression pieces for the lowest possible overhead once compiled.