[BUGFIX] Fix compiling of boolean nodes with hardcoded string comparisons
The way nodes were being converted was causing expressions which (for example) compared a variable against a hardcoded string to fail. By processing the expression parts before analysing them this problem is fixed. Also implemented is a slight performance improvement in the AbstractConditionViewHelper which allows the class to use closures only, lowers the size of compiled code and prevents a few method calls.