[TASK] Evict Exception throwing in ViewHelperVariableContainer
Exceptions are a very invasive way of dealing with the fact that some variable does not exist in what is essentially a big storage array. We remove these thrown Exceptions, replacing them with NULL returns and allowing a default argument for the `get()` method which will be returned if the variable does not exist.