[FEATURE] Allow default content on optional f:render
Whenever `f:render` is called and the `optional` flag is set to `TRUE`, rendering a missing section results in an empty output. Instead of rendering this empty output, a new attribute `default` (mixed) is added and can be filled with a fallback-type default value. Alternatively, the tag content can be used to define this default value like so many other content/attribute-flexible ViewHelpers.