[BUGFIX] Remove ``render()`` method from ``AbstractConditionViewHelper``
This removes the render() method from the AbstractViewHelper
class in order to prevent PHP errors when a subclass overrides the
method with a different signature (i.e. including parameters).

Note:
This was lost when re-applying https://github.com/NamelessCoder/TYPO3.Fluid/pull/18
to this repository.
This change also adds the ``AbstractViewHelperTest`` changes that got
lost.