[BUGFIX] Let f:cycle use "as" argument as scope of cycled variable
This change allows f:cycle to be nested (see linked issue) - to work with multiple cycled variables in the same loop. In essence it identifies the instance of "values" using the value of the "as" argument and uses ViewHelperVariableContainer for storing the index.

Also changes the ViewHelper to not store the values but rather just store the index and read out the appropriate value based on that, from the provided values in arguments.

References: #5