[FEATURE] Allow namespaces to be extended
Allows:

```php
// given class My\Package\ViewHelpers\RenderViewHelper exists and REPLACES f:render
$view->getViewHelperResolver()->extendNamespace('f', 'My\\Package\\ViewHelpers');
$class = $view->getViewHelperResolver()->resolveViewHelperClassName('f', 'render');
// "My\\Package\\ViewHelpers\\RenderViewHelper"

// given class My\Package\OtherViewHelper exists, gets ADDED to "f:" namespace
$class = $view->getViewHelperResolver()->resolveViewHelperClassName('f', 'other');
// "My\\Package\\ViewHelpers\\OtherViewHelper"
```

And in templates allows all ViewHelpers from "My\Package\ViewHelpers" to be used as part of "f:".

The feature is achieved through allowing each registered PHP namespace string to become an array and the resolve class method to support both strings and arrays, walking the arrays backwards (for "overlay" style behavior) when it detects multiple PHP namespace targets.