[FEATURE] RenderingContext as implementation API
This change:

* Wraps all essential context-like classes into RenderingContext
* Removes those same context-like classes from various instances which now receive a RenderingContext

Using the RenderingContext is ideal since it is available both during parsing and once templates are compiled, and can be replaced at any time without invalidating compiled templates.

The change breaks code which depends on the following usages:

* getTemplatePaths() on View
* getViewHelperResolver() on View
* getViewHelperInvoker() on ViewHelperResolver

Every such usage can be replaced by $renderingContext->getTemplatePaths(), $renderingContext->getViewHelperResolver() and so on. Included examples have been updated to reflect this.