[FEATURE] Replace escaping by f:format.htmlspecialchars with EscapingNode
Rather than use the f:format.htmlspecialchars ViewHelper via a ViewHelperNode, involving resolving and everything else implied by rendering a ViewHelper, a new node type - EscapingNode - is implemented. The node is added in the same way as the old escaping would do, but evaluates as and compiles to a single `htmlspecialchars` function call.