[FEATURE] Allow f:alias to create VariableProviders using "src" argument
This adds three new arguments to f:alias:

* `src` which can contain data which can be read by a VariableProvider
* `name` which can contain the name of a "container" variable in which values from the Provider will be placed
* `type` which allows manually defining the type

Currently, the one VariableProvider type which makes sense to use here is the `JsonVariableProvider`.

Examples:

```xml
<!-- given "file.json" contains {"name": "John", "phone": "1-800-FLUID"} -->
<f:alias src="/path/to/file.json" as="json">
  {json.name} at {json.phone}
  <!-- outputs: "John at 1-800-FLUID" -->
</f:alias>
```

The "src" argument can any stream wrapper compatible URI.

In addition the "JSONVariableProvider" has been renamed to "JsonVariableProvider" to be able to correctly assume the name of this and future VariableProviders to be supported.