[BUGFIX] Fixes for namespace resolving
Before this change, ignoring a namespace with `{namespace ignoredname}` was not respected due to casting NULL to an array. And support restored for namespaces added as PHP namespace in xmlns notation, including without the "ViewHelpers" suffix which will now get appended automatically.

This change also removes a redundant argument on ViewHelperResolver and adds a specific method to check if namespace is ignored.