[BUGFIX] Correct handling of disabled compiler in relation to cache
This change makes sure that when caching is disabled, the template compiler will flush the cache for any existing entries (allowing for example dynamically flushing cache by custom TTL from a ViewHelper's compile method). Until this change, a previously cached template could never be excluded from compiling, which resulted in bad behavior when combined with the fact that the template caching could possibly already have begun at the time when a ViewHelper disables the compiler.