[BUGFIX] Compiled condition never reaches "else" if passed as argument
When rendering a template with a piece of template code like:

```xml
{f:if(condition: variable, then: 'yes', else: 'no')}
<f:if condition="{variable}" then="yes" else="no" />
```

Would never execute the "else" statement once the ViewHelper was compiled regardless of input arguments. Used with child nodes the compiled result works perfectly.

Problem is in renderStatic which does not consider "else" passed directly in arguments. The condition is extended to also consider this way of passing the argument.