[BUGFIX] Invalid array object access
Accessing an object is bound to the existence of a method called
"exists()" which leads to an invalid array access on objects actually
implementing that method without the intention to be accessed as array.

Using the ArrayAccess interface using the method "offsetExists()"
offers a more specific approach on using an array accessor however.