[BUGFIX] Assign RenderingContext during parsing
Some third party ViewHelpers may depend on the presence of a rendering context while parsing - although it technically is an incorrect expectation to have the context available at this point, making it available allows third parties to access various secondary aspects of the rendering context.

Specifically, the lack of this parse-time rendering context setting affects Widgets used in TYPO3 CMS when the template is uncompilable.

Followup for https://github.com/TYPO3Fluid/Fluid/pull/77 which removed this line as it was evidently redundant. Line restored now along with a comment linking to this pull request.