[FEATURE] (Re-)Introduce escaping modifiers functionality
You can now disable the escaping interceptor by adding
the following line anywhere in your template:

{escapingEnabled=false}

This disables the output and children escaping for all view helpers
during the complete rendering. You cannot add this more than once
in your template.

It is also possible to set

{escapingEnabled=true}

but this currently is a noop, as esacping is enabled by default
in all contexts (rendering formats).

Fixes #99