[BUGFIX] Do not sort by key when receiving paths arrays
During the decoupling a decision was made to not use
array keys for the priority of template paths - instead,
the order of the array should determine the lookup order.

However, the TemplatePaths class of TYPO3Fluid had
received a sorting behavior that was intended for the
TemplatePaths class of TYPO3 CMS. This patch removes
that sorting behavior.

This fix must be followed up by a patch for TYPO3 CMS.
suggested strategy there is to override the setter methods
for paths and enforce the sorting there.

References: #104
References: https://forge.typo3.org/issues/75862
References: https://review.typo3.org/#/c/48279/