[BUGFIX] Detect/allow number-prefixed variables in array parts
This change fixes an issue where a template variable named for example "123variable" would render correctly when referenced as `{123variable}` but if used in an array part like for example `<f:alias map="{newVar: 123variable}">` the variable would not be resolved. Instead, a numeric value of `123` would be cast by the parser. Loosening up the expression which required array part variable references to start with a letter solves the issue.