[FEATURE] Provide dedicated ViewHelperCompiler and Traits (#113)
This change adds a dedicated ViewHelperCompiler to cover frequent use cases in ViewHelpers:

* Compiling directly with static function call (to renderStatic - or other methods on other classes)
* Compiling with static function call and using one of the provided arguments as "content" or calling renderChildren if that argument is not passed.

Both use cases support calling a static method on *any* class - not just ViewHelpers, or on another ViewHelper - if that method conforms to the signature of the renderStatic method defined in ViewHelperInterface.

Both use cases are then implemented in ViewHelpers of Fluid where possible, which improves performance of the following ViewHelpers:

* f:format.cdata
* f:format.printf
* f:alias
* f:comment
* f:count
* f:for
* f:groupedFor
* f:or
* f:spaceless

Other ViewHelpers are, with one exception (f:cycle) already compilable in the most efficient way. The cycle ViewHelper still requires a custom compile method that will probably be so adapted it cannot be made generic enough to merit putting it in a Trait and the ViewHelperCompiler.