[FEATURE] Introduce FailedCompilingState
This change introduces a new type of ParsingState
(ParsedTemplateInterface implementation) which is to
be used when compiling fails.

Currently this new type is not implemented anywhere
but is planned for the coming cache warmup feature as
a way to provide better feedback about failed compiles.
The class includes properties to carry both a failure
description and suggestions to mitigate the problem;
the intention being that such solutions shall be possible
to suggest when throwing Exceptions from ViewHelpers.

In the long run it is the intention to use this parsing state
whenever compiling fails, thus allowing things like debug
inspections to report problems and potential solutions,
while preserving the original ParsingState as a production
alternative that doesn’t catch and report compiling issues.