[BUGFIX] Fix compiling for renderStatic with content argument
This change fixes an issue where the mentioned compiling
method would produce invalid PHP syntax if the content
argument could not be resolved.

Furthermore, uses the protected property for content
argument name on the Trait for storing the resolved name.

Also adds examples and test cases that validate this
compiling method.