[BUGFIX] Make content-argument Trait generate Closure
The default closure generated by ViewHelpers implementing
the CompileWithContentArgumentAndRenderStatic Trait is
not appropriate for rendering the content argument when the
template is not compiled.

This also allows dropping the (statically called, non-API)
method which compiles ViewHelpers differently when they
use a content argument. Now, the standard method to call
renderStatic will receive a Closure fitting the arguments.

Close: #209