[FEATURE] Allow command-line Fluid to include third-party bootstrap
The support added by this patch is fully described at
https://forge.typo3.org/issues/78665, but a short
summary would be that a new "bootstrap" argument is
added to the Fluid CLI command which when used,
includes that bootstrap file (and does nothing else).

Pending the linked changed in TYPO3 this allows Fluid
to operate on command line using every context and
setting configured in the site it is installed in.