[BUGFIX] Avoid iteration data in f:for without "iteration" argument
Skip counting and other operations for iteration data when
the "iteration" argument is not set. Counting for example
QueryResult can be rather expensive so avoiding it whenever
possible makes good sense.

Also switches some `!== null` checks in favor of `isset()`,
which is done for consistency and to avoid potential e_notice.

Close: #243