[BUGFIX] Allow renderSection() to be called directly
Refactors AbstractTemplateView slightly so that
the parsed template reading code can be called
from both render() and renderSection(), causing
the template to be loaded correctly.

Without this patch it is not possible to call
renderSection() via the View's API without getting
an error that no parsed template exists.

Also adds a public method on TemplateCompiler
to reset the current processing state.