[TASK] Refactor condition ViewHelper logic
This patch makes a few but breaking changes to the
way AbstractConditionViewHelper and implementations
function. The following is changed:

* Method `render` is now delared on AbstractConditionViewHelper
   and removed from IfViewHelper. Implementers of the
   AbstractConditionViewHelper class no longer need to
   implement a `render` method - `evaluateCondition` is
   enough; and if necessary, additional/other VH arguments.
* Method `evaluateCondition` is now declared abstract
   on AbstractConditionViewHelper, meaning that any
   implementations must implement this method.
* Argument `condition` is no longer automatically added
   by subclassing AbstractConditionViewHelper. This
   argument is instead added by implementations, which
   must also call `parent::initializeArguments` to gain "then"
   and "else" argument support. The "then" and "else"
   arguments are in addition now optional (but `f:then` and
   `f:else` node support is forced).
* Implementations of AbstractConditionViewHelper can now
   implement any argument as condition (file for "file exists",
   haystack and needle for "string contains", and so on).

The change is breaking to any implementations of the
AbstractConditionViewHelper class which 1) do not declare
an `evaluateCondition` method and/or 2) overrides arguments
initialisation without calling `parent::initializeArguments` or
finally, 3) assumes that `condition` argument gets passed but
does not explicitly declare it as argument.

Subclasses of IfViewHelper are not prone to this problem.
Should you need a quick fix for your custom condition VH then
simply switching parent class to IfViewHelper should in most
cases make your subclass work again.