[FEATURE] Synchronise xmlns extraction with TYPO3 CMS
Due to slight differences in how Fluid detects xmlns declarations
in templates, TYPO3 CMS has had to deliver not one, but two
separate methods of detecting and registering namespaces.

This patch synchronises the behavior and removes the need for
TYPO3 CMS to process the template for namespaces.

There is one main difference introduced by this patch:

All xmlns declarations are now stripped from the tag; and the tag
itself is stripped away if `data-namespace-typo3-fluid="true"`
is set on the tag.