[FEATURE] Add escaping modifier pre-processor (#266)
This patch adds support for disabling escaping throughout an entire template file. This is done by using:

```
{escaping off}
```

In a template, layout or partial. When toggled off it causes the template parser to skip escaping of all outputs from ViewHelpers, variable outputs etc.

For legacy reasons the feature is added with backwards compatibility for the original escaping modifier which was introduced in Flow. Therefore, the following values are all equally possible and mean the same:

```
{escaping off}
{escapingEnabled off}
{escaping false}
{escaping=false}
{escaping=off}
{escaping = off}
{escapingEnabled = off}
{escapingEnabled = false}
{escapingEnabled off}
{escapingEnabled false}
```

While it is possible to set a `true` value for the modifier, this simply gets ignored (because escaping is the default).