[FEATURE] Add ErrorHandler with TolerantErrorHandler implementation
This patch adds the possibility to replace the error handling
that Fluid normally does, which until now was simply to throw
exceptions, break the entire rendering and leave it up to the
integration how to handle the exception.

By replacing the ErrorHandlerInterface implementation it is
possible to make an ErrorHandler which instead of throwing
exceptions, returns a simple string or silently logs the problem.

The default implementation is enabled by default and will still
throw exceptions like normal. An alternative implementation is
also included, which converts normally fatal errors to simple
strings that are then output in the template where the error is.

Usage:

```
$view->getRenderingContext()->setErrorHandler(new \TYPO3Fluid\Fluid\Core\ErrorHandler\TolerantErrorHandler());
```