[FEATURE] Add TagGenerating trait and TagViewHelper
Provides a Trait as replacement for tag based ViewHelpers,
which can be implemented by any ViewHelper, along with
an implementation that allows generating arbitrary tags while
utilising the features of the TagBuilder.

Using the trait is slightly different from using the tag based
abstract ViewHelper class in that all the actual rendering
takes place through a static API method createTag() which
receives arguments and attributes separately and returns
an instance of TagBuilder which is then rendered automatically.

Also refactors TagBuilder to allow chaining and support for
closures as tag content.